[wip][build] Failed attempt at reducing compilation times
 - GCC ~1min 15s
 - Clang ~ 26s!

Moving initialization of some static variables like FixedHuffmanCoding
and blockfinder/DynamicHuffman into their own translations units did
improve as good as nothing!

Avoiding the 16-bit reversedBits LUT shaves off ~10s for GCC but that is
negligible...

Note that total compilation times actually increase when not using mold!

The traces generated with Clang with -ftime-trace compilation option
only showed reverseBits with ~2s as some outlier but the rest is simply
too homogenously distributed amongst all functions. Half the time is
spent in the optimization phase. It might help to explicitly instantiate
as much as possible, e.g., the Huffman Coding implementations. But,
because a lot is used in constexpr contexts, it is difficult to do.