[performance] BlockFinder: Use WalkTreeCompressedLUT with a single LUT for +10% speed: 64 -> 72 MB/s
Parameterize WalkTreeCompressedLUT chunk count and use limit of 7,
i.e., a single LUT for the whole precode histogram.

This also reduces the binary size from 6.5 MiB to 2.5 MiB
because the histogram LUT is dynamically initialized and even if it was
not, the LUT size would have reduced from 4 MiB -> ~1.9 MB.

See printCompressedPrecodeFrequenciesLUTSizes<7U, 512U> in
testPrecodeCheck.cpp.

The next commit will add precomputed LUTs to fix the startup times.
This commit degraded a simple rapidgzip --help from ~5ms to 1s!