[performance] Remove 3 more bits for almost free by rearranging the histogram bin order
Unfortunately, this only reduces LUT size by about 2x.
It seems that there are that many fewer duplicate chunks now.
Maybe we can rearrange the inner bins even further until we find the
optimal size...?

At least the bandwidth recovered and is now almost as fast as the
original WalkTreeLUT with all 7 bins included in the LUT.